update from dev pull origin dev, modify CommentServiceTest

+++ src/main/java/com/example/codePicasso/domain/exchange/service/ExchangeService.java
@@ -2,7 +2,7 @@ package com.example.codePicasso.domain.exchange.service;
 
 import com.example.codePicasso.domain.exchange.dto.request.ExchangeRequest;
 import com.example.codePicasso.domain.exchange.dto.request.MyExchangeRequest;
-import com.example.codePicasso.domain.exchange.dto.request.PutExchangeRequest;
+import com.example.codePicasso.domain.exchange.dto.request.PutMyExchangeRequest;
 import com.example.codePicasso.domain.exchange.dto.response.ExchangeResponse;
 import com.example.codePicasso.domain.exchange.dto.response.MyExchangeResponse;
 import com.example.codePicasso.domain.exchange.entity.Exchange;
@@ -92,9 +92,20 @@ public class ExchangeService {
     public void deleteExchange(Long exchangeId, Long userId) {
         Exchange exchange = exchangeConnector.findById(exchangeId);
 
+        if (exchange.getUser() == null) {
+            throw new NotFoundException(ErrorCode.USER_NOT_FOUND);
+        }
+
         if (!exchange.getUser().getId().equals(userId)) {
             throw new NotFoundException(ErrorCode.USER_NOT_FOUND);
         }
+
+        if (exchange.isCompleted()) {
+            log.debug("이미 완료된 거래, 예외 발생 예정.");
+            throw new InvalidRequestException(ErrorCode.ALREADY_IN_COMPLETED);
+        }
+
+        exchange.completed();
 
         exchangeConnector.deleteById(exchangeId);
     }
@@ -109,7 +120,7 @@ public class ExchangeService {
         Exchange exchange = exchangeConnector.findById(exchangeId);
         User user = userConnector.findById(userId);
 
-        if (myExchangeConnector.existsByExchangeIdAndUserId(exchangeId, userId)) {
+        if (myExchangeConnector.existByExchangeIdAndUserId(exchangeId, userId)) {
             throw new DuplicateException(ErrorCode.DUPLICATE);
         }
 
@@ -142,18 +153,18 @@ public class ExchangeService {
 
     // 거래 승인/거절/취소하기 (200 OK)
     @Transactional
-    public void decisionMyExchange(Long myExchangeId, Long userId, PutExchangeRequest putExchangeRequest) {
+    public void decisionMyExchange(Long myExchangeId, Long userId, PutMyExchangeRequest putExchangeRequest) {
         MyExchange myExchange = myExchangeConnector.findById(myExchangeId);
 
         if (!myExchange.getExchange().getUser().getId().equals(userId)) {
             throw new NotFoundException(ErrorCode.USER_NOT_FOUND);
         }
 
-        if (myExchange.getExchange().getStatusType() == StatusType.COMPLETED) {
+        if (myExchange.getStatusType() == StatusType.COMPLETED) {
             throw new InvalidRequestException(ErrorCode.ALREADY_IN_COMPLETED);
         }
 
-        myExchange.getExchange().changeStatus(putExchangeRequest.statusType());
+        myExchange.changeStatus(putExchangeRequest.statusType());
         myExchangeConnector.save(myExchange);
 
         if (putExchangeRequest.statusType() == StatusType.COMPLETED) {
@@ -163,7 +174,7 @@ public class ExchangeService {
 
     /// --- ↑ MyExchange ---
 
-    /// --- ↓ Redis ---
+    /// --- ↓ dblock ---
 
     @Transactional
     public void completeExchange(Long exchangeId) {
@@ -178,7 +189,7 @@ public class ExchangeService {
         boolean isBuy = exchange.getTradeType() == TradeType.BUY;
 
         try {
-            exchangeRankingService.increaseTradeCount(exchange.getGame().getGameTitle(), isBuy);
+            exchangeRankingService.increaseTradeCount(exchange.getGame().getId(), exchange.getGame().getGameTitle(), isBuy);
         } catch (Exception e) {
             log.error("Redis 업데이트 실패: gameTitle={}, isBuy={}", exchange.getGame().getGameTitle(), isBuy, e);
             throw new DataAccessException(ErrorCode.TRADE_RANKING_UPDATE_FAILED) {
@@ -188,5 +199,5 @@ public class ExchangeService {
         }
     }
 
-    /// --- ↑ Redis ---
+    /// --- ↑ dblock ---
 }
